Minor refactor for reliability
