+++ reports/duplicates/requirements/base.txt
@@ -1,3 +1,4 @@
 amclient
 pandas==0.24.2
+metsrw==0.2.0
